+++ CopilotKit/packages/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @copilotkit/runtime-client-gql
 
+## 1.8.7-next.0
+
+### Patch Changes
+
+- 8b8474f: - feat: add image input support with multi-model compatibility, pasting, and UX improvements
+  - @copilotkit/shared@1.8.7-next.0
+
+## 1.8.6
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.6
+
 ## 1.8.6-next.0
 
 ### Patch Changes