+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @copilotkit/shared
 
+## 1.8.7-next.0
+
+## 1.8.6
+
 ## 1.8.6-next.0
 
 ## 1.8.5